+++ eras/shelley/impl/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Add `runImpTestM`, `runImpTestM_`, `evalImpTestM` and `execImpTestM`
 * Add instance `Example (a -> ImpTestM era ())`, which allows use of `Arbitrary`
+* Add `getNextEpochCommitteeMembers` to `EraGov` typeclass, with a default empty implementation
 
 ## 1.8.0.0